+++ CSK_Module_ImagePlayer/scripts/Sensors/ImagePlayer/helper/checkAPIs.lua
@@ -1,0 +1,42 @@
+---@diagnostic disable: undefined-global, redundant-parameter, missing-parameter
+
+-- Load all relevant APIs for this module
+--**************************************************************************
+
+local availableAPIs = {}
+
+local function loadAPIs()
+  CSK_ImagePlayer = require 'API.CSK_ImagePlayer'
+
+  Container = require 'API.Container'
+  DateTime = require 'API.DateTime'
+  Engine = require 'API.Engine'
+  Log = require 'API.Log'
+  Log.Handler = require 'API.Log.Handler'
+  Log.SharedLogger = require 'API.Log.SharedLogger'
+  Object = require 'API.Object'
+  Timer = require 'API.Timer'
+  View = require 'API.View'
+
+  -- Check if related CSK modules are available to be used
+  local appList = Engine.listApps()
+  for i = 1, #appList do
+    if appList[i] == 'CSK_Module_PersistentData' then
+      CSK_PersistentData = require 'API.CSK_PersistentData'
+    elseif appList[i] == 'CSK_Module_UserManagement' then
+      CSK_UserManagement = require 'API.CSK_UserManagement'
+    end
+  end
+end
+
+local function loadSpecificAPIs()
+  Image = require 'API.Image'
+  Image.Provider = {}
+  Image.Provider.Directory = require 'API.Image.Provider.Directory'
+end
+
+availableAPIs.default = xpcall(loadAPIs, debug.traceback) -- TRUE if all default APIs were loaded correctly
+availableAPIs.specific = xpcall(loadSpecificAPIs, debug.traceback) -- TRUE if all specific APIs were loaded correctly
+
+return availableAPIs
+--**************************************************************************